+++ novawallet/Common/Services/RemoteSubscription/Substrate/ParachainStakingAccountSubscriptionService.swift
@@ -2,54 +2,40 @@ import Foundation
 import SubstrateSdk
 import Operation_iOS
 
-protocol ParachainStakingAccountSubscriptionServiceProtocol {
+extension ParachainStaking {
+    final class AccountSubscriptionService: RemoteSubscriptionService {
+        private static let storagePaths: [StorageCodingPath] = [
+            ParachainStaking.delegatorStatePath
+        ]
+    }
+}
+
+extension ParachainStaking.AccountSubscriptionService: StakingRemoteAccountSubscriptionServiceProtocol {
     func attachToAccountData(
         for chainAccountId: ChainAccountId,
         queue: DispatchQueue?,
         closure: RemoteSubscriptionClosure?
-    ) -> UUID?
+    ) -> UUID? {
+        attachToAccountDataWithStoragePaths(
+            Self.storagePaths,
+            chainAccountId: chainAccountId,
+            queue: queue,
+            closure: closure
+        )
+    }
 
     func detachFromAccountData(
         for subscriptionId: UUID,
         chainAccountId: ChainAccountId,
         queue: DispatchQueue?,
         closure: RemoteSubscriptionClosure?
-    )
-}
-
-extension ParachainStaking {
-    final class AccountSubscriptionService: RemoteSubscriptionService,
-        ParachainStakingAccountSubscriptionServiceProtocol {
-        private static let storagePaths: [StorageCodingPath] = [
-            ParachainStaking.delegatorStatePath
-        ]
-
-        func attachToAccountData(
-            for chainAccountId: ChainAccountId,
-            queue: DispatchQueue?,
-            closure: RemoteSubscriptionClosure?
-        ) -> UUID? {
-            attachToAccountDataWithStoragePaths(
-                Self.storagePaths,
-                chainAccountId: chainAccountId,
-                queue: queue,
-                closure: closure
-            )
-        }
-
-        func detachFromAccountData(
-            for subscriptionId: UUID,
-            chainAccountId: ChainAccountId,
-            queue: DispatchQueue?,
-            closure: RemoteSubscriptionClosure?
-        ) {
-            detachFromAccountDataStoragePaths(
-                Self.storagePaths,
-                subscriptionId: subscriptionId,
-                chainAccountId: chainAccountId,
-                queue: queue,
-                closure: closure
-            )
-        }
+    ) {
+        detachFromAccountDataStoragePaths(
+            Self.storagePaths,
+            subscriptionId: subscriptionId,
+            chainAccountId: chainAccountId,
+            queue: queue,
+            closure: closure
+        )
     }
 }
